AWSteria: Host program reads elf file (specified on command line) directly
The host-side software is in src_Host_Side, and
developer_designs/cl_BSV_WindSoC/software/runtime is now a symlink to this
directory.
   "make" accepts the following targets:
o test_bsim (the default): builds an executable to run as peer of a bluesim
simulation of the hardware;
o test_vsim:  builds an executable to run as peer of a verilator simulation of
the hardware;
o test_aws: build an executable to connect to a dut running in an AWS FPGA.

The program can also be run as the host side of an xsim cosimulation; this is
run from developer_designs/cl_BSV_WindSoC/verif/scripts in the way explained
in the AWS documentation.  Since this is used only rarely, and to avoid
excessive complications, in this case an elf file is *not* specified on the
command line; instead the program expects to find a Mem.hex file for the
program to be run, in
developer_designs/cl_BSV_WindSoC/verif/sim/vivado/test_c.

The main polling loop of the program has not changed from previous versions;
so drafts of changes (e.g. to provide for virtio) should still be valid.

The different flows are controlled by macros AWS_BSIM, AWS_VSIM, AWS_FPGA and
SV_TEST.  The last of these (indicating cosimulation) is provided by the
AWS/Xilinx infrastructure for the xsim target.  Most of the conditionals
controlled by these macros are in the library file AWS_BS_Lib.c; they occur in
the main test.c file only when necessary.