Hoist SoC out of src_Testbench_AWS and drop _AWS from src_Testbench_AWS
Calling the SoC part of the testbench is confusing, as the term
testbench is generally reserved for simulation. Instead, pull it out
into its own src_SoC directory at the top level.

Also be clearer about what's platform-specific and what's generic across
AWS and a VCU118-based platform by renaming src_Testbench_AWS to just
src_Testbench and adding an AWS subdirectory to src_SoC for the boot
ROM, which needs to embed a device tree. Though ideally this would be
generated during the build from parameterisable source (and thus be
located in the build directory, not the source directory).