Add configuration for a VCU118 framework, Garnet
Garnet[1] is a cut-down reimplementation of the AWS F1 shell for the
VCU118. This adds all the same configurations as exist for AWS.

There is a lot of duplication between all these files. Some thought
should probably go into how to reduce that, perhaps by auto-generating
the Makefiles on demand as supported by Flute rather than committing
many similar copies.

The boot ROM is currently almost certainly wrong, as it is just a copy
of the AWS one. The original source for this should be added to the
repository and the bsvi file built as part of the build rather than
committed as source. This both makes it a less mysterious entity and
allows for easy modification (including parameterisation).

The names for many of the files/modules/functions/etc are now slightly
misleading, with many using AWS or AWS-specific terms in their names.
Ideally better names would be used, but that's a rather invasive change
that would be primarily cosmetic.

No script (or SystemVerilog wrapper) is provided yet to build a
bitstream, but that would be rather futile without support for multiple
clock domains in the SoC.

[1] https://github.com/CTSRD-CHERI/garnet